requirements-dev.txt
-r requirements.txt
pytest==7.2.0
pyinstaller>=6.0
pillow>=10.0
